integrations/docling-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aspose-slides-docling-backend"
dynamic = ["version"]
dependencies = [
  "docling",
  "aspose.slides>=25.7.0"
]

[tool.hatch.version]
path = "src/aspose_slides_docling_backend/__about__.py"
